+++ components/shared/include/shared.h
@@ -6,7 +6,11 @@ typedef struct {
     float temperature;
     float humidity;
     float pressure;
+    uint8_t moisture;
     uint16_t lux;
+    uint16_t nitrogen;
+    uint16_t phosphorous;
+    uint16_t potassium;
 } shared_data_t;
 
 extern shared_data_t shared_data;