+++ main/main.c
@@ -65,6 +65,26 @@ void app_main(void)
             APP_CPU_NUM
     );
 
+//    xTaskCreatePinnedToCore(
+//            sen0605_task,
+//            "sen0605",
+//            configMINIMAL_STACK_SIZE * 8,
+//            NULL,
+//            10,
+//            NULL,
+//            APP_CPU_NUM
+//    );
+
+    xTaskCreatePinnedToCore(
+            moisture_task,
+            "moisture",
+            configMINIMAL_STACK_SIZE * 8,
+            NULL,
+            10,
+            NULL,
+            APP_CPU_NUM
+    );
+
     // Start the NimBLE host task
     nimble_port_freertos_init(bleprph_host_task);